Allow being served from a non-`/` path
This is relevant for e.g. github pages or if served through another
reverse proxy.
The issue lies in building up paths after first load. If the user
reloads the page or is redirected back to a deep-link, the router must
route back to the expected place (and the webserver must _not_ just try
to find a static file in a non-existent place in the file system!).

It relies on the fact that github expects a `404.html` file to exist
which replaces the standard "skull 404" page. If this is instead the
normal angular app, everything works as expected.